Add simple SVG export helper
